+++ src/Drawing.cpp
@@ -88,7 +88,7 @@ void DrawCircle(
     GLfloat x,
     GLfloat y,
     GLfloat radius,
-    GLfloat sides,
+    GLuint sides,
     GLfloat lineWidth
 )
 {
@@ -125,8 +125,8 @@ void DrawTexture(
         GLfloat texBottom = (GLfloat)imgHeight / (GLfloat)texHeight;
         GLfloat texLeft = 0.f;
         GLfloat texRight = (GLfloat)imgWidth / (GLfloat)texWidth;
-        GLfloat quadWidth = imgWidth;
-        GLfloat quadHeight = imgHeight;
+        GLfloat quadWidth = (GLfloat) imgWidth;
+        GLfloat quadHeight = (GLfloat) imgHeight;
 
         if(clip != NULL)
         {
@@ -182,7 +182,7 @@ void DrawText(
 
     GLfloat curX = x;
     GLfloat curY = y;
-    for(int i = 0; i < text.size(); ++i)
+    for(int i = 0; i < (int)text.size(); ++i)
     {
         if(text[i] == '\n' || text[i] == '\r')
         {
